% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_pd.R
\name{funDist}
\alias{funDist}
\title{Distance function}
\usage{
funDist(grid, x, y, metric, p = NULL, lambda = NULL, phi = NULL)
}
\arguments{
\item{grid}{the grid (of length \code{T}) over which the two curves are defined.}

\item{x}{a vector containing the first curve.}

\item{y}{a vector containing the second curve.}

\item{metric}{the chosen distance to be used: \code{"L2"} for the classical L2-distance, \code{"trunc"} for the truncated Mahalanobis semi-distance, \code{"mahalanobis"} for the generalized Mahalanobis distance}

\item{p}{a positive numeric value containing the parameter of the regularizing function for the generalized Mahalanobis distance.}

\item{lambda}{a vector containing the eigenvalues of the functional data from which \code{"x"} and \code{"y"} are extracted.}

\item{phi}{a matrix containing the eigenfunctions of the functional data from which \code{"x"} and \code{"y"} are extracted.}
}
\value{
The function returns a numeric value indicating the distance between the two curves.
}
\description{
This function allows you to compute the distance between two curves with the chosen metric.
}
\examples{
# Define parameters
n <- 50
P <- 100
K <- 150

# Grid of the functional dataset
t <- seq( 0, 1, length.out = P )

# Define the means and the parameters to use in the simulation
# with the Karhunen - Loève expansion
m1 <- t^2 * ( 1 - t )

lambda <- rep( 0, K )
theta <- matrix( 0, K, P )
for ( k in 1:K ) {
  lambda[k] <- 1 / ( k + 1 )^2
  if ( k\%\%2 == 0 )
    theta[k, ] <- sqrt( 2 ) * sin( k * pi * t )
  else if ( k\%\%2 != 0 && k != 1 )
    theta[k, ] <- sqrt( 2 ) * cos( ( k - 1 ) * pi * t )
  else
    theta[k, ] <- rep( 1, P )
}

# Simulate the functional data
z <- simulate_KL( t, n, m1, rho = lambda, theta = theta )

# Extract two rows of the functional data
x <- z$data[[1]][1, ]
y <- z$data[[1]][2, ]

d <- funDist( t, x, y, metric = "L2" )

}
\references{
Ghiglietti A., Ieva F., Paganoni A. M. (2017). Statistical inference for stochastic processes:
Two-sample hypothesis tests, \emph{Journal of Statistical Planning and Inference}, 180:49-68.

Ghiglietti A., Paganoni A. M. (2017). Exact tests for the means of gaussian stochastic processes.
\emph{Statics & Probability Letters}, 131:102--107.
}
