% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hhat}
\alias{hhat}
\title{Regularizing h function}
\usage{
hhat(p, lambda)
}
\arguments{
\item{p}{a positive numeric value containing the parameter of the regularizing function for the generalized Mahalanobis distance.}

\item{lambda}{a vector containing the eigenvalues of the covariance matrix of the functional data from which \code{"x"} and \code{"y"} are extracted.}
}
\value{
The function returns a vector of the values of the regularizing function for the generalized Mahalanobis distance.
}
\description{
This function computes the values of the regularizing function for the generalized Mahalanobis distance.
}
\examples{
# Define the parameters of the function
p <- 10^3
K <- 150
lambda <- rep( 0, K )
for ( k in 1:K ) {
  lambda[k] <- 1 / ( k + 1 )^2
}

h <- hhat( p, lambda )
}
