% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{integral}
\alias{integral}
\title{Integral function}
\usage{
integral(grid, values)
}
\arguments{
\item{grid}{a vector}

\item{values}{a vector}
}
\value{
The function returns a numeric value indicating the value of the integral under the curve.
}
\description{
This function computes the integral of a function defined on a grid of values.
}
\examples{
grid <- seq( 0, 1, length.out = 1000 )
values <- grid^2
I <- integral( grid, values )
}
