% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funData.R
\name{plot.funData}
\alias{plot.funData}
\title{A method to plot \code{funData} objects}
\usage{
\method{plot}{funData}(x, ...)
}
\arguments{
\item{x}{the univariate functional dataset in form of \code{funData} object.}

\item{...}{additional graphical parameters to be used in plotting functions}
}
\description{
This function performs the plot of a functional univariate dataset stored in
an object of class \code{funData}.
}
\examples{

N = 20

# Define parameters
n <- 50
P <- 100
K <- 150

# Grid of the functional dataset
t <- seq( 0, 1, length.out = P )

# Define the means and the parameters to use in the simulation
# with the Karhunen - Loève expansion
m1 <- t^2 * ( 1 - t )
m2 <- t * ( 1 - t )^2

lambda <- rep( 0, K )
theta <- matrix( 0, K, P )
for ( k in 1:K) {
  lambda[k] <- 1 / ( k + 1 )^2
  if ( k\%\%2 == 0 )
    theta[k, ] <- sqrt( 2 ) * sin( k * pi * t )
  else if ( k\%\%2 != 0 && k != 1 )
    theta[k, ] <- sqrt( 2 ) * cos( ( k - 1 ) * pi * t )
  else
    theta[k, ] <- rep( 1, P )
}
# Simulate the functional data
x1 <- simulate_KL( t, n, m1, rho = lambda, theta = theta )
x2 <- simulate_KL( t, n, m2, rho = lambda, theta = theta )

FD <- funData( t, list( as.matrix( x1$data[[1]] ), as.matrix( x2$data[[1]] ) ) )

plot(FD)

}
\seealso{
\code{\link{funData}}
}
